backend/tools/requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asyncpg==0.27.0
pydantic==1.10.9

